+++ branches/branch-playground/sim/socket.cpp
@@ -180,7 +180,7 @@ void ClientSocket::read_ready()
         if (readn == 0)
           break;
         readBuffer().setWritePos(readBuffer().writePos() + readn);
-        if(readBuffer().writePos() < readBuffer().size())
+        if(readBuffer().writePos() < (unsigned)readBuffer().size())
           break;
         if (m_notify)
             m_notify->packet_ready();
